pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-relay"
version = "0.1.0"
description = "MCP middleware relay and transport utilities"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.13"
authors = [
    { name = "Todd Firsich", github = "BoilerToad" }
]
keywords = ["mcp", "middleware", "relay", "llm", "transport"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]

dependencies = [
    "mcp>=1.26.0",
    "ollama>=0.4.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
postgres = ["asyncpg>=0.29.0"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[project.urls]
Repository = "https://github.com/BoilerToad/mcp-relay"
Issues = "https://github.com/BoilerToad/mcp-relay/issues"

[tool.hatch.build.targets.wheel]
packages = ["mcp_relay"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "integration: requires Ollama running locally (deselect with '-m not integration')",
    "slow: tests that take more than 5 seconds",
]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.13"
strict = true
